moved reading and writing functions to BasePacket class
