+++ try1/Core/Networking/Packets/BasePacket.cs
@@ -1,19 +1,49 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ChatChimpClient.Core.PacketHandlers
+namespace ChatChimpClient.Core.Networking.Packets
 {
-    public class PackageReader
+    public class BasePacket
     {
-        private BinaryReader reader { get; set; }
-        public PackageReader(byte[] data)
+        BinaryReader reader { get; set; }
+        BinaryWriter writer { get; set; }
+        public BasePacket(byte[] data)
         {
             MemoryStream ms = new MemoryStream(data);
             reader = new BinaryReader(ms);
+            writer = new BinaryWriter(ms);
         }
+
+        // ---------------- WRITER FUNCTIONS ----------------
+
+        public void writeByte(byte b)
+            => writer.Write(b);
+
+        public void writeUshort(ushort number)
+        {
+            byte[] toWrite = BitConverter.GetBytes(number);
+            for ( int x = 0; x < 2; x++ )
+            {
+                writer.Write(toWrite[x]);
+            }
+        }
+
+        public void writeInt(int number)
+        {
+            writer.Write7BitEncodedInt(number);
+        }
+
+        public void writeString(string letters)
+        {
+            int msgLen = Encoding.UTF8.GetBytes(letters).Length;
+            writeInt(msgLen);
+            writer.Write(letters);
+        }
+
+        // ------------ READING FUNCTIONS --------------
 
         public int readIntBytes()
         {
@@ -53,4 +83,4 @@ namespace ChatChimpClient.Core.PacketHandlers
             reader.BaseStream.Seek(1, SeekOrigin.Current);
         }
     }
-}//
+}
